Don't attempt to set guifont
Managing fonts across machines is a huge pain, and different fonts may
be more appropriate for different machines.  So instead, load a local
~/.gvimrc-local if it exists, the same way ~/.vimrc-local is respected.
Arguably, there's now an argument for dropping the .gvimrc link
altogether and just keep it an entirely machine-specific resource.  I'll
stick to this for now as it's easier to expand upon, but maybe that
argument will be revisited.